+++ test/node/download-dht-magnet.js
@@ -7,7 +7,7 @@ var test = require('tape')
 var WebTorrent = require('../../')
 
 test('Download using DHT (via magnet uri)', function (t) {
-  t.plan(11)
+  t.plan(12)
 
   var dhtServer = new DHT({ bootstrap: false })
 
@@ -25,6 +25,10 @@ test('Download using DHT (via magnet uri)', function (t) {
       client1 = new WebTorrent({
         tracker: false,
         dht: { bootstrap: '127.0.0.1:' + dhtServer.address().port, host: networkAddress.ipv4() }
+      })
+
+      client1.dht.on('listening', function () {
+        t.equal(client1.dhtPort, client1.dht.address().port)
       })
 
       client1.on('error', function (err) { t.fail(err) })